Release 1.4.2
* MAC Factory & Hallways

The MAC factory has been completely redone, with new graphics and mechanics.  It now requries 4 cannon rounds and 4 batteries to create a single slug, depleting its entire internal storage.  It can only store a single slug inside itself.

Hallways are like omni-directional moving walkways that accelerate crew to a modest speed.  Not working atm.

Tungsten carbide armor is now backwards-compatible with ship files from when it was part of the WWII mod.

* EMP icon

The EMP Cannon's icon is now the appropriate size in the Ship Designer